Restructuring, altering imports, removing axios util.
