--- src/axios.js
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-const instance = axios.create({
-  baseURL: `${process.env.REACT_APP_API_URI}`,
-  responseType: 'json',
-});
-
-export default instance;
