--- src/components/pages/Home/RenderHomePage.js
@@ -1,34 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-import { Button } from '../../common';
-
-function RenderHomePage(props) {
-  const { userInfo, authService } = props;
-  return (
-    <div>
-      <h1>Hi {userInfo.name} Welcome to Labs Basic SPA</h1>
-      <div>
-        <p>
-          This is an example of a common example of how we'd like for you to
-          approach components.
-        </p>
-        <p>
-          <Link to="/profile-list">Profiles Example</Link>
-        </p>
-        <p>
-          <Link to="/example-list">Example List of Items</Link>
-        </p>
-        <p>
-          <Link to="/datavis">Data Visualizations Example</Link>
-        </p>
-        <p>
-          <Button
-            handleClick={() => authService.logout()}
-            buttonText="Logout"
-          />
-        </p>
-      </div>
-    </div>
-  );
-}
-export default RenderHomePage;
