+++ src/styles/theme-overrides.js
@@ -1,6 +1,6 @@
 module.exports = {
-  '@primary-color': '#ec3944',
-  '@link-color': ' #ec3944',
+  '@primary-color': '#6199b5',
+  '@link-color': '#6199b5',
   '@success-color': '#52c41a',
   '@warning-color': '#faad14',
   '@error-color': '#f5222d',
@@ -9,7 +9,7 @@ module.exports = {
   '@text-color': 'rgba(0, 0, 0, 0.65)',
   '@text-color-secondary': 'rgba(0, 0, 0, 0.45)',
   '@disabled-color': 'rgba(0, 0, 0, 0.25)',
-  '@border-radius-base': '2px',
+  '@border-radius-base': '4px',
   '@border-color-base': '#d9d9d9',
   '@box-shadow-base':
     '0 3px 6px -4px rgba(0, 0, 0, 0.12), 0 6px 16px 0 rgba(0, 0, 0, 0.08), 0 9px 28px 8px rgba(0, 0, 0, 0.05)',
